+++ src/commands/expressions/index.ts
@@ -1,3 +1,4 @@
 export * from "./condition.js";
 export * from "./key-condition.js";
 export * from "./operands.js";
+export * from "./update.js";